game: shops — und der Builder findet einen Fehler in guests, den nur er finden konnte
Das Modul steht: 6.170 Zeilen, eine Warteschlange je Laden, eine Kasse je Tresen,
Lager, Lieferungen, Einnahmen, Absagen, Personalbedarf, und
`find(need, x, z, cash)`, sortiert nach Gehweg plus Wartezeit in Parkminuten.
Auf der Hauptseite ein parametrischer Gebäudegenerator mit fünf Baukörpern, ein
prozeduraler 4×2-PBR-Atlas, Thin Instances und ein Nachtlicht-Pool.

Die Kosten sind pro TYP und nicht pro Laden: zwölf Ladentypen sind 25 Meshes und
19.094 Dreiecke, ein Park mit acht Burgerbuden kostet also so viel wie einer mit
einem. Im Demo-Park sind sieben Läden 1,4 % des Draw-Call-Budgets und 0,83 % der
Dreiecke des Bildes. Sim-Tick 0,0007 ms. Der 51-Prüfungen-Selbsttest hängt jetzt
als `test:game-shops` in `pnpm test:game`.

Der wichtigere Teil dieses Commits ist aber ein Fehler in `guests`, den dieser
Builder finden musste, weil er erklären wollte, warum niemand etwas kauft:
**die Bedürfnisse stiegen bei keiner Geschwindigkeit, die ein Spieler wählen
kann.** `d.needs` war ein Uint8Array, beschrieben durch `Math.round`, und der
Anstieg je Tick beträgt 0,0217 Einheiten bei Geschwindigkeit 1, 0,065 bei 3,
0,108 bei 5 und 0,217 bei 10 — jeder einzelne rundet auf null. Er bewegt sich
erst bei 100×, und das ist genau die Geschwindigkeit, mit der der Soak läuft:
der einzige Test, der dieses Modul je bewegt hat, lief bei der einen
Geschwindigkeit, bei der der Fehler unsichtbar ist.

Gemessen vorher: 934 von 944 Gästen stehen untätig ohne Ziel nach drei
Parkstunden, mittlerer Hunger 24 gegen ein `urgentAt` von 170. Nach dem
Verbreitern der Spalte auf Float32: 115,6 / 116,0 / 115,8 bei den
Geschwindigkeiten 1, 3 und 5 nach 300 Parkminuten — über alle drei identisch,
was eine zeitbasierte Abnahme sein muss. Ein vor der Verbreiterung geschriebener
Spielstand lädt weiter, die alte Byte-Spalte wird beim Laden aufgeweitet.

Dazu drei Dinge vom Integrator. Der Demo-Park stellt seine Läden jetzt hin —
sieben von acht geplanten Plätzen, ausgewählt nach BEDÜRFNIS gegen die Registry
statt nach Pack-Id, bei x = ±11, weil die Laternen der Hauptachse bei ±6,2 und
ihre Linden bei ±8 stehen und der Dienstradius des Wegegraphen 14 m beträgt. Der
achte Platz bleibt leer, weil die gebündelten Packs nur einen Souvenirladen
haben: der Code lässt einen Platz lieber leer, als einen Laden zu erfinden.

Und das Harness starb unter SwiftShader zuverlässig an
„Resulting promise was garbage collected", sobald `--step=N` auf eine Szene traf,
die den Render-Loop ins Stocken bringt — vier von vier Läufen bei `overview`, und
mit den Läden im Demo-Park auch bei `ground`. `waitForFunction` legt ein
langlebiges Promise in die Seite; jetzt wird von node aus in kurzen Aufrufen
gepollt.

Offen und ausdrücklich nicht halb gemacht: es gibt ZWEI Modelle fürs Kaufen.
`guests/sim.ts` liest Laden-Entities selbst aus der Welt, schlägt ihre
Definitionen in der Registry nach und fährt eine eigene Warteschlange mit
eigenem Preis — daher die Einnahmen im Bild, die Kasse geht von 2.500.000 auf
2.500.063 —, während die Warteschlangen-, Kassen- und Lagermodellierung von
`shops` überhaupt keinen Aufrufer hat. Ihr eigener Bericht sagt es: kein Bild
zeigt eine Schlange an einem Tresen, weil nichts ansteht. Das umzuhängen ist die
nächste Runde für `guests` und kein Flicken.

pnpm test:game grün (16 Prüfungen), tsc/eslint/prettier sauber, 0 Konsolenfehler.

+++ lib/game/guests/decide.ts
@@ -53,7 +53,7 @@ export interface DecisionContext {
   needs: NeedModel;
   archetype: GuestArchetypeDef;
   /** Guest's need row, 0..255. */
-  levels: Uint8Array;
+  levels: Float32Array;
   /** Offset of the guest's row in `levels`. */
   base: number;
   x: number;
